engine: extract esm/tours/overlay.js
